Extract sanity check to its own test class
